In all commands error handling is implemented and now copy command is required to implement

+++ deletedir.cpp
@@ -9,12 +9,19 @@ vector<string> deleteDir(vector<string> &commandSplit)
     {
     	remove_argument= new char[commandSplit[i].length()+1];
     	strcpy(remove_argument,commandSplit[i].c_str());
-    	status = rmdir(remove_argument);
-	    if(status!=0)
-	    {
-	        deletefilesfolders.push_back(commandSplit[i]);
-	        deleteall(commandSplit[i]);
-	    }
+    	DIR *d;
+    	d=opendir(remove_argument);
+    	if(d)
+    	{
+    		status = rmdir(remove_argument);
+		    if(status!=0)
+		    {
+		        deletefilesfolders.push_back(commandSplit[i]);
+		        deleteall(commandSplit[i]);
+		    }
+    	}
+    	else
+    		cout<<endl<<"No such directory exist"<<endl;  
     }
     /*for(unsigned i=0;i<deletefilesfolders.size();i++)
     {
@@ -35,12 +42,18 @@ void deleteall(string rm)
 	strcpy(rm1,rm.c_str());
 	char *rm2;
 	int total = scandir(rm1,&namelist, NULL,alphasort);
+
 	for(i=0;i<total;i++)
 	{
 		path=rm+"/"+namelist[i]->d_name;
 		rm2=new char[path.length()+1];
 		strcpy(rm2,path.c_str());
-		stat(rm2,&statObj);
+		//stat(rm2,&statObj);
+		if(stat(rm2,&statObj) < 0)   
+	    {
+	        cout<<endl<<"No such directory exist"<<endl;
+	        //return 1;
+	    } 
 		if(S_ISDIR(statObj.st_mode) && string(namelist[i]->d_name)!="." && string(namelist[i]->d_name)!="..")
 		{
 			//int f=current.find_last_of("/\\");
